+++ packages/core/.storybook/main.ts
@@ -73,6 +73,14 @@ export default {
       },
       define: {
         "process.env.NODE_ENV": JSON.stringify("development")
+      },
+      optimizeDeps: {
+        include: ["storybook-addon-playground"]
+      },
+      build: {
+        commonjsOptions: {
+          exclude: ["storybook-addon-playground"]
+        }
       }
     });
   }