+++ packages/core/src/components/Avatar/Avatar.tsx
@@ -15,6 +15,7 @@ import ClickableWrapper from "../Clickable/ClickableWrapper";
 import { type VibeComponentProps, withStaticPropsWithoutForwardRef } from "../../types";
 import { type SubIcon } from "@vibe/icon";
 import styles from "./Avatar.module.scss";
+import { ComponentVibeId } from "../../tests/constants";
 
 export interface AvatarProps extends VibeComponentProps {
   /**
@@ -221,6 +222,7 @@ const Avatar = ({
       data-testid={dataTestId || getTestId(ComponentDefaultTestId.AVATAR, id)}
       className={cx(styles.avatar, styles[size], className)}
       style={sizeStyle}
+      data-vibe={ComponentVibeId.AVATAR}
     >
       <ClickableWrapper
         isClickable={!!onClick}