+++ packages/core/src/components/Combobox/Combobox.tsx
@@ -27,6 +27,7 @@ import styles from "./Combobox.module.scss";
 import { type ComboboxSizes } from "./Combobox.types";
 import type IconButton from "../IconButton/IconButton";
 import type MenuButton from "../MenuButton/MenuButton";
+import { ComponentVibeId } from "../../tests/constants";
 import { type SubIcon } from "@vibe/icon";
 
 export interface ComboboxProps extends VibeComponentProps {
@@ -368,6 +369,7 @@ const Combobox = forwardRef(
         id={id}
         data-testid={dataTestId || getTestId(ComponentDefaultTestId.COMBOBOX, id)}
         ellipsis={false}
+        data-vibe={ComponentVibeId.COMBOBOX}
       >
         <div className={styles.comboboxList} style={{ maxHeight: optionsListHeight }}>
           <Search