+++ packages/core/src/components/Counter/Counter.tsx
@@ -15,6 +15,7 @@ import {
 import { type CounterColor, type CounterSize, type CounterType } from "./Counter.types";
 import { type VibeComponentProps, withStaticPropsWithoutForwardRef } from "../../types";
 import styles from "./Counter.module.scss";
+import { ComponentVibeId } from "../../tests/constants";
 
 export interface CounterProps extends VibeComponentProps {
   /**
@@ -139,6 +140,7 @@ const Counter = ({
       aria-label={`${ariaLabel} ${countText}`}
       aria-labelledby={ariaLabeledBy}
       onMouseDown={onMouseDown}
+      data-vibe={ComponentVibeId.COUNTER}
     >
       <div className={classNames} aria-label={countText} ref={ref}>
         {noAnimation ? (
@@ -154,10 +156,8 @@ const Counter = ({
                 exit: styles.fadeExit,
                 exitActive: styles.fadeExitActive
               }}
-              // @ts-expect-error @definitelyTyped typings expecting a single parameter for some reason when the function passed here is called with two parameters
-              // See https://github.com/reactjs/react-transition-group/blob/c89f807067b32eea6f68fd6c622190d88ced82e2/src/Transition.js#L522-L534
-              addEndListener={(node: HTMLElement, done: () => void) => {
-                node.addEventListener("transitionend", done, false);
+              addEndListener={done => {
+                nodeRef.current?.addEventListener("transitionend", done, false);
               }}
             >
               <span