+++ packages/core/src/components/Modal/Modal/Modal.tsx
@@ -4,7 +4,7 @@ import { RemoveScroll } from "react-remove-scroll";
 import FocusLock from "react-focus-lock";
 import { motion, AnimatePresence } from "framer-motion";
 import { getTestId } from "../../../tests/test-ids-utils";
-import { ComponentDefaultTestId } from "../../../tests/constants";
+import { ComponentDefaultTestId, ComponentVibeId } from "../../../tests/constants";
 import styles from "./Modal.module.scss";
 import { type ModalProps } from "./Modal.types";
 import ModalTopActions from "../ModalTopActions/ModalTopActions";
@@ -162,6 +162,7 @@ const Modal = forwardRef(
                         )}
                         id={id}
                         data-testid={dataTestId || getTestId(ComponentDefaultTestId.MODAL_NEXT, id)}
+                        data-vibe={ComponentVibeId.MODAL}
                         role="dialog"
                         aria-modal
                         aria-labelledby={ariaLabelledby || titleId}