+++ packages/core/CHANGELOG.md
@@ -3,6 +3,68 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.61.1](https://github.com/mondaycom/vibe/compare/@vibe/core@3.61.0...@vibe/core@3.61.1) (2025-08-25)
+
+
+### Bug Fixes
+
+* tabs screen reader ([#3066](https://github.com/mondaycom/vibe/issues/3066)) ([9e39770](https://github.com/mondaycom/vibe/commit/9e39770629d925892694982686967db90e59b004))
+
+
+
+
+
+# [3.61.0](https://github.com/mondaycom/vibe/compare/@vibe/core@3.60.0...@vibe/core@3.61.0) (2025-08-20)
+
+
+### Features
+
+* IME support for EditableHeading, EditableText ([#3052](https://github.com/mondaycom/vibe/issues/3052)) ([7074f42](https://github.com/mondaycom/vibe/commit/7074f42e983e377ee1f1f0aa6b88178920fe08fb))
+* **Link:** add `style` prop for inline styling ([#3063](https://github.com/mondaycom/vibe/issues/3063)) ([48df2c5](https://github.com/mondaycom/vibe/commit/48df2c5afd3a647718e3c7025b29fe5da3724c5c))
+* **Tipseen:** add onShow event ([#3064](https://github.com/mondaycom/vibe/issues/3064)) ([d5954e7](https://github.com/mondaycom/vibe/commit/d5954e75cb79e37f65dbb02e88ad2415683792e9))
+
+
+### Reverts
+
+* "feat(Tipseen): add onShow event" ([#3065](https://github.com/mondaycom/vibe/issues/3065)) ([2df6c8b](https://github.com/mondaycom/vibe/commit/2df6c8b70c21fbcc1c55aee052b6c9eecdcde306))
+
+
+
+
+
+# [3.60.0](https://github.com/mondaycom/vibe/compare/@vibe/core@3.59.0...@vibe/core@3.60.0) (2025-08-14)
+
+
+### Features
+
+* **Table:** use custom scrollbar ([#3032](https://github.com/mondaycom/vibe/issues/3032)) ([60962e6](https://github.com/mondaycom/vibe/commit/60962e68c94a80c184617132236fdf5969b86e34))
+
+
+
+
+
+# [3.59.0](https://github.com/mondaycom/vibe/compare/@vibe/core@3.58.4...@vibe/core@3.59.0) (2025-08-13)
+
+
+### Features
+
+* **Toggle:** add noSpacing prop to remove horizontal spacing ([#3048](https://github.com/mondaycom/vibe/issues/3048)) ([e0c6732](https://github.com/mondaycom/vibe/commit/e0c6732bbdda1268f453efe375caeba054a15a1e))
+
+
+
+
+
+## [3.58.4](https://github.com/mondaycom/vibe/compare/@vibe/core@3.58.3...@vibe/core@3.58.4) (2025-08-12)
+
+
+### Bug Fixes
+
+* **Counter:** fix usage of CSSTransition#addEndListener with nodeRef ([#3047](https://github.com/mondaycom/vibe/issues/3047)) ([a73b8e9](https://github.com/mondaycom/vibe/commit/a73b8e972144051449b1f47a2c7ee1b5e2d560d1))
+
+
+
+
+
 ## [3.58.3](https://github.com/mondaycom/vibe/compare/@vibe/core@3.58.2...@vibe/core@3.58.3) (2025-08-12)